requirements.txt
numpy
pillow
cupy_cuda12x
opencv_python_headless
nvidia_dali_cuda120